Bump version to 1.0.0
## Why major/breaking changes version?
This version of the app will solely relying on MKVToolNix to handle metadata and manipulation

Using FFMPEG may requires remuxing the file, which is not ideal for large files, IO intensive, and time consuming.

Additionally, we moved when to import  Matplotlib's `font_manager` to be on-demand instead. This is due to how slow the app will load on first startup and it really not needed when user prefer to use SRT only.